Add hint injection
When a write io arrives, it's matched against a list of hints according
to offset and size. This hint can affect the tier the write will be
routed to later. If no hint is found, the write is suspended until such
hint arrives. The hint can be injected using a new ioctl on tiercontrol.